+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## HEAD (Unreleased)
 
+### Improvements
+
 -   Support helm v3 `include-crds` argument. (https://github.com/pulumi/pulumi-kubernetes/pull/1102)
+-   Bump python requests version dependency. (https://github.com/pulumi/pulumi-kubernetes/pull/1121)
 
 ## 2.1.1 (May 8, 2020)